Fix in navigation with waze app
the url of the waze application contained an error. in property 'll' the latitude and longitude of the starting point was being used, when in fact it should use the target latitude and longitude. So when the app opened it would direct you to where you already are. When you changed the property to the destination point everything worked as expected.