Add option to always include Google Maps
The use of Google Maps is the default for so many users. This change adds an option for users to always be able to select Google Maps as their choice, even if the Google Maps app is not installed on their device.

When this option is enabled, it will use the url with https:// maps.google.com, which will open in the Google Maps app, if installed, otherwise it will open in Safari.

There would be no change for Android users.